Arraylist works at last

+++ src/model/animal.java
@@ -5,7 +5,7 @@ public class animal extends Toy {
 	char size;
 	
 	
-	public animal(int serial, String name, String brand, float price, int availableCount, int ageRating, String material, char size) {
+	public animal(long serial, String name, String brand, float price, int availableCount, int ageRating, String material, char size) {
 		super(serial, name, brand, price, availableCount, ageRating);
 		this.material = material;
 		this.size = size;
